separated out the components in Register.jsx
